add batt_p instead of batt_v
batt_p is the remaining battery as a percentage, which is an
easier way to think about than an arbitrary scaled millivolt
value. While the code is not taking action based on percent
yet, it may in the future, including warnings and stuff.

In addition, this basically shuts down when the battery
is sufficiently depleted that the ADC starts to incorrectly
scale the voltages as the reference voltage sags. This
could lead to incorrectly reporting that the pilot light
is out, when really, it is just a low battery.

Signed-off-by: Vernon Mauery <vernon@mauery.org>